requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
tensorflow-cpu
gunicorn
werkzeug
